refactor: 패키지명을 도메인에 맞게 변경
