+++ src/main/java/lotto/rank/Rank.java
@@ -1,4 +1,4 @@
-package lotto.match;
+package lotto.rank;
 
 public enum Rank {
     FIRST(6, 2_000_000_000),
